Tidied main js to remove toasts and create separate js file which is then imported to main
