+++ src/main.js
@@ -1,4 +1,12 @@
 import './assets/main.css'
+import {
+  infoToast,
+  successToast,
+  errorToast,
+  Toast,
+  goodRequest,
+  badRequest
+} from './js/toast.js'
 
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
@@ -6,9 +14,6 @@ import piniaPersist from 'pinia-plugin-persist'
 import App from './App.vue'
 import router from './router'
 import axios from 'axios'
-import Toast from "vue-toastification";
-import "vue-toastification/dist/index.css";
-import { useToast } from 'vue-toastification'
 
 const pinia = createPinia()
 pinia.use(piniaPersist)
@@ -32,70 +37,11 @@ app.config.globalProperties.$userStore = userStore();
 app.config.globalProperties.$gameStore = gameStore();
 app.config.globalProperties.$loading = loadingStore();
 app.config.globalProperties.$clueStore = clueStore();
-app.config.globalProperties.$toast = useToast();
-
-const toast = useToast(); 
-const errorToast = (message) => toast.error(message, {
-  position: "top-right",
-  timeout: 3000,
-  closeOnClick: true,
-  pauseOnFocusLoss: true,
-  pauseOnHover: true,
-  showCloseButtonOnHover: false,
-  hideProgressBar: true,
-  closeButton: "button",
-  icon: true,
-  rtl: false
-})
-const successToast = (message) => toast.success(message, {
-  position: "top-right",
-  timeout: 3000,
-  closeOnClick: true,
-  pauseOnFocusLoss: true,
-  pauseOnHover: true,
-  showCloseButtonOnHover: false,
-  hideProgressBar: true,
-  closeButton: "button",
-  icon: true,
-  rtl: false
-})
-const infoToast = (message) => toast.info(message, {
-  position: "top-right",
-  timeout: 3000,
-  closeOnClick: true,
-  pauseOnFocusLoss: true,
-  pauseOnHover: true,
-  showCloseButtonOnHover: false,
-  hideProgressBar: true,
-  closeButton: "button",
-  icon: true,
-  rtl: false
-})
-
 app.config.globalProperties.$error= errorToast;
 app.config.globalProperties.$success= successToast;
 app.config.globalProperties.$info= infoToast;
-
-app.config.globalProperties.$badRequest = (axioserror) => { 
-  const status = axioserror?.response?.status || 500 
-  const message = axioserror?.response?.data?.message || 'Something went wrong...'
-  console.log(axioserror)
-  if (status > 499 ){ 
-    errorToast('Server Error');
-  } 
-  if (status < 500 && status > 399) { 
-    errorToast(message);
-  }
-}
-
-app.config.globalProperties.$goodRequest = (axiosmessage) => { 
-  const status = axiosmessage?.response?.status || 200 
-  const message = axiosmessage?.response?.data?.message || 'Success!'
-  console.log(axiosmessage)
-  if (status > 199 && status > 300){ 
-    successToast('Success!');
-  } 
-}
+app.config.globalProperties.$goodRequest = goodRequest;
+app.config.globalProperties.$badRequest = badRequest;
 
 const socket = io.connect('http://localhost:5000');
 function connectSocket(socket) {
